+++ truck/src/main/java/com/manager/truck/truck/usecase/user/GetUserByIdUseCase.java
@@ -1,4 +1,4 @@
-package com.manager.truck.truck.usecase;
+package com.manager.truck.truck.usecase.user;
 
 import com.manager.truck.truck.domain.User;
 import com.manager.truck.truck.domain.dto.response.UserResponse;
@@ -7,14 +7,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
-public class DeleteUserById {
+public class GetUserByIdUseCase {
     @Autowired
     private UserService userService;
 
-    public void execute(Long id){
-        User user= userService.findById(id);
-        user.setState(false);
-
-        userService.update(user);
+    public UserResponse execute(Long id){
+        User builduser = userService.findById(id);
+        return new UserResponse(builduser);
     }
 }
