+++ truck/src/main/java/com/manager/truck/truck/usecase/user/UserSaveUseCase.java
@@ -1,9 +1,8 @@
-package com.manager.truck.truck.usecase;
+package com.manager.truck.truck.usecase.user;
 
 import com.manager.truck.truck.domain.User;
 import com.manager.truck.truck.domain.dto.request.UserRequest;
 import com.manager.truck.truck.domain.dto.response.UserResponse;
-import com.manager.truck.truck.repository.UserRepository;
 import com.manager.truck.truck.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
